install packages before to replace trans, if not installed

+++ install-fedora.sh
@@ -41,6 +41,11 @@ function install_trans {
     echo -ne "installing ${1} (${2}) "
     for i in "${TRANS[@]}"; do
         IFS=':' read -a ADDR <<< "${i//"LOCALE"/${LANG_CODE}}"
+        dnf provides ${ADDR[1]} &> /dev/null
+        if [ $? -ne 0 ]; then
+            echo -ne "[pakage ${ADDR[1]} was not installed. installing...] "
+            dnf install -y `dnf repoquery -f ${ADDR[1]}` &> /dev/null && echo "${GREEN}[ OK ]${NC}" || exit 1
+        fi
         rm -f ${ADDR[1]} && msgfmt ${BASE_PATH}/${1}-${2}/${ADDR[0]} -o ${ADDR[1]} && echo "${GREEN}[ OK ]${NC}" || echo "${RED}[ FAIL ]${NC}"
     done
 }
